Define a GNU-compatible getline() implementation, in case we don't have getline() available.

+++ changes/fergus.henderson/k9j6/trunk/src/util.c
@@ -837,6 +837,6 @@ ssize_t getline(char **lineptr, size_t *n, FILE *stream) {
     *lineptr = buffer;
     *n = size;
 
-    return bytes_read == 0 ? -1 : bytes_read;
+    return bytes_read == 0 ? -1 : (ssize_t) bytes_read;
 }
 #endif
